task-guacamole-restricted
Set up a special task for packages with restrictive (i.e., commercial)
licenses: start with a list of packages we might want in the images and
filter this down to those that the license whitelist permits. Emit
warning for any of these packages that are not whitelisted, but do not
include them in the image.