mx: patch for more efficient creation of MxComboBox
mx_combo_box_append_text() and friends destroy and reconstruct the entire
menu on each call, which is extremely inefficient (i.e., to construct a
menu with 10 entries, 9 intermediate menus with 1, 2, ... 9, items are
constructed and destroyed). Even on a fairly powerfull platform such as Atom,
a box with 50 items takes seconds to construct, 150 entries whole 45s!

This patch adds mx_combo_box_populate() which allows to append a bunch of
items to the combo box in a single go and only construct the menu once.