Work around broken Yocto beagleboard.conf
Yocto provides a beagleboard.conf which is no compatible with meta-ti (and
consdiring the very limited support Yocto provides for it, is pretty useless).
Unfortunately, because meta-yocto *preppends* itself to the BBPATH and meta-ti
appends itself, there is no way in which meta-yocto could be combined with
meta-ti and use the correct beagleboard.conf (i.e., the one from meta-ti).

So symlink beagleboard.conf from meta-ti under meta-guacamayo.