udev: fix SD card rootfs getting mounted under /media
Currently if the rootfs is on an SD card, the Yocto mount.sh script will mount
it again under /media; this is undesirable at best. In the Guacamayo case,
all removable devices, i.e., devices mounted under /media, get indexed by
tracker, so having the rootfs mounted there pretty much kills the system.

The modified mount.sh script checks, after a successful mount, whether a well
known file under the new mount point is the same file as the equivalent on the
root device (we use /etc/udev/udev.conf); if it does we promptly again umount
the /media mount point.

The current solution is a woraround rather than a proper fix, as ideally we
would want to avoid mounting the fs in the first place; something like
blkid could be used here, but we don't have that.