task-guacamayo: remove explicit alsa-plugins build-dependency
This was there because to get Pulse and ALSA talking nicely you need
alsa-plugins, but as nothing build-depends on it it won't get built for the
image.

Now that we have images without Pulse (mediaserver), remove this dependency and
ensure that it's listed in all tasks that actually need it.