guacamayo-session-common: add .this-is-root marker to the root of the rootfs
We need some reliable way to tell if a fs mounted under media is actually
the rootfs, since Yocto messes us about here: if the rootfs is on an SD card,
it will be mounted again under /media. Ideally a fix to the mount handling
is needed so that the duplicate mount is not created, but it is surprisingly
hard to detect/handle this correctly, because on live images, the rootfs in
fact needs to be mounted under /media during the early booting stage.

Our main problem is tracker indexing the rootfs, because of a patch to the
removable media detection we apply in the first place. So put a marker into
the rootfs and tweak the patch.