gst-ffmpeg.bbappend: For RPi revert back to the configure options from Denzil
The configure options in the forthcoming Danny, which specify exact --cc and
--as options, flush out a bug in the internal copy of libav, using assembly
code that is not compatible with the RPi CPU. For now, revert back to the
configure options from Denzil.

The same bug affects the standalone version of libav, and meta-raspberrypi
has a patch for the assembly and some extra configure paramers. It is might
worth porting this over to the gst-ffmpeg package.