guacamayo-sesion-egl (RPi): allow the user group to execute shutdown
There is no power button on the RPi and simply pulling the plug is bad for
the SD card. So, allow the user to execute the shutdown command, which we
can then run from MEX when a remote Power button is pressed.

This is a very crude solution, but will do for now and this specific case.