pulseaudio: provide specific pa config for the nuc
For some reason the HDMI card in the nuc is not detected using udev, for
now statically load pa alsa module for both of devices the HDMI card
provides.

The other consequence of the udev detection not working for this card is
that we only get audio on the HDMI1 output; needs a proper fix.